requirements.txt
Flask==2.2.5
opencv-python==4.8.0.76
face_recognition==1.3.0
numpy==1.24.0
Pillow==9.5.0
pyserial==3.5
pandas==2.1.0